requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
yfinance==0.2.33
fredapi==0.5.1
plotly==5.18.0
streamlit==1.29.0
python-dateutil==2.8.2
pytz==2023.3
scipy==1.11.4
